Koopman embedding using KAN network
The Koopman embedding employs a KAN network in both the encoder and decoder modules, with both networks having identical architectures. One notable extension is the use of an embedding size of 64 instead of 32, nearly doubling the number of parameters to 105,645. Additionally, there is a change in how the model handles the evaluation of reconstruction loss. In the original implementation of transformer physics, the evaluation was performed between (x(t) and x(t+1), which is not the correct approach and the loss was only calculated for the first time series.