Implementation of KAN network
This implementation of the KAN network is based on the efficient KAN repository, with modifications to make the model more verbose in printing its architecture. It is easier to implement in PyTorch compared to the PyKAN implementation, which includes multiple other functions that complicate training with other neural networks using nn.Sequential in PyTorch.