A mixup between KAN and MLP
Instead of using only MLP or KAN in the encoder and decoder, this architecture incorporates an MLP layer followed by a Tanh layer to transform the input of both the encoder and decoder. This helps adjust the input to the desired range for the KAN network and captures simple non-linear patterns.